add all non-alphanumeric characters to replace
Still having issues with getting some of the non-alphanumeric characters. So,
rather than hit and miss, I figured it would be best to just search and replace
them all. If there are still pastebins that come through that don't work, I'll
have to look at them at an individual basis.