Optimize sha512crypt()
By taking the conditional flow control out of the loop, the execution is
greatly increased. This is done by regognizing that LCM(2,3,7)=42
(intentional?). Thus, what gets added to digest "c" is repeated every 42 times
through the loop. So, by building up a list of all possible 42 permutations,
then using that to build the digests, we can reduce a great about of work.

50% performance increase!